Add a CloudFormation template for public AWS CI (#58)
* Add a CloudFormation template for public AWS CI

* Add documentation on test stack creation

This required setting up the general Documenter infrastructure as well.
Note that linking docstrings into the documentation and all of that has
not been done as part of this commit.
The text for the section on test stack setup was taken from that in the
CloudWatchLogs package, written by Eric Davies. Their authorship as been
included as part of this commit for this reason.

Co-authored-by: Eric Davies <iamed2@gmail.com>

* Add a configuration file for bors

* Only run Travis from specific branches

In particular, build on master, tags, and bors branches.